Convert congregations page to react using Grommet.
Initially it looks good and is a lot simpler to get it all wired
than using Fabric. The API is clearer and simpler. Layout is a breeze.
Some things are just too big, but a custom theme should work well.

The only remaining page is the languages page and it can probably be removed now. With that I can turn this into a SPA and simplify all the UI code.